mijnvisie/requirements.txt
fastapi[standard]
pysondb-v2
passlib[bcrypt]
pyjwt
pydantic
pydantic_settings